requirements.txt
aiogram>=3.0.0
fastapi
uvicorn
python-dotenv